PWN-6888 - move history logic to a common HistoryListView.kt and HistoryListViewPresenter.kt and use it in HistoryFragment.kt and TokenHistoryFragment.kt
