+++ app/src/main/java/org/p2p/wallet/history/ui/history/HistoryFragment.kt
@@ -1,31 +1,21 @@
 package org.p2p.wallet.history.ui.history
 
-import androidx.core.view.isVisible
-import androidx.recyclerview.widget.LinearLayoutManager
 import android.os.Bundle
 import android.view.View
 import org.koin.android.ext.android.inject
 import timber.log.Timber
-import org.p2p.core.glide.GlideManager
 import org.p2p.core.token.Token
-import org.p2p.uikit.utils.attachAdapter
-import org.p2p.uikit.utils.recycler.RoundedDecoration
 import org.p2p.wallet.R
 import org.p2p.wallet.common.mvp.BaseMvpFragment
-import org.p2p.wallet.common.ui.recycler.EndlessScrollListener
-import org.p2p.wallet.common.ui.recycler.PagingState
 import org.p2p.wallet.databinding.FragmentHistoryBinding
-import org.p2p.wallet.history.model.HistoryItem
 import org.p2p.wallet.history.model.HistoryTransaction
 import org.p2p.wallet.history.model.TransactionDetailsLaunchState
 import org.p2p.wallet.history.ui.detailsbottomsheet.HistoryTransactionDetailsBottomSheetFragment
-import org.p2p.wallet.history.ui.token.adapter.HistoryAdapter
 import org.p2p.wallet.moonpay.ui.new.NewBuyFragment
 import org.p2p.wallet.moonpay.ui.transaction.SellTransactionDetailsBottomSheet
 import org.p2p.wallet.receive.solana.ReceiveSolanaFragment
 import org.p2p.wallet.sell.ui.lock.SellTransactionViewDetails
 import org.p2p.wallet.utils.replaceFragment
-import org.p2p.wallet.utils.unsafeLazy
 import org.p2p.wallet.utils.viewbinding.viewBinding
 
 class HistoryFragment :
@@ -39,95 +29,23 @@ class HistoryFragment :
     override val presenter: HistoryContract.Presenter by inject()
     private val binding: FragmentHistoryBinding by viewBinding()
 
-    private val glideManager: GlideManager by inject()
-    private val adapter: HistoryAdapter by unsafeLazy {
-        HistoryAdapter(
-            glideManager = glideManager,
-            onHistoryItemClicked = presenter::onItemClicked,
-            onRetryClicked = presenter::loadNextHistoryPage
-        )
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        with(binding.layoutHistoryList) {
-
-            errorStateLayout.buttonRetry.setOnClickListener {
-                presenter.refreshHistory()
-            }
-            val scrollListener = EndlessScrollListener(
-                layoutManager = historyRecyclerView.layoutManager as LinearLayoutManager,
-                loadNextPage = { presenter.loadNextHistoryPage() }
-            )
-
-            historyRecyclerView.addOnScrollListener(scrollListener)
-            historyRecyclerView.attachAdapter(adapter)
-            historyRecyclerView.addItemDecoration(RoundedDecoration(16f))
-            refreshLayout.setOnRefreshListener {
-                presenter.refreshHistory()
-                scrollListener.reset()
-            }
-
-            emptyStateLayout.buttonBuy.setOnClickListener {
-                presenter.onBuyClicked()
-            }
-            emptyStateLayout.buttonReceive.setOnClickListener {
-                replaceFragment(ReceiveSolanaFragment.create(token = null))
-            }
-        }
+        binding.layoutHistoryList.bind(
+            onBuyClicked = presenter::onBuyClicked,
+            onReceiveClicked = { replaceFragment(ReceiveSolanaFragment.create(token = null)) },
+            onHistoryItemClicked = presenter::onItemClicked
+        )
 
         listenForSellTransactionDialogDismiss()
         lifecycle.addObserver(presenter)
+        binding.layoutHistoryList.addObserver(lifecycle)
     }
 
     private fun listenForSellTransactionDialogDismiss() {
         childFragmentManager.setFragmentResultListener(
             SellTransactionDetailsBottomSheet.REQUEST_KEY_DISMISSED, this
-        ) { _, _ -> presenter.loadHistory() }
-    }
-
-    override fun showPagingState(state: PagingState) {
-        adapter.setPagingState(state)
-        with(binding.layoutHistoryList) {
-            when (state) {
-                is PagingState.InitialLoading -> {
-                    shimmerView.root.isVisible = true
-                    refreshLayout.isVisible = false
-                }
-                is PagingState.Idle -> {
-                    shimmerView.root.isVisible = false
-                    refreshLayout.isVisible = true
-                    errorStateLayout.root.isVisible = false
-                    emptyStateLayout.root.isVisible = adapter.isEmpty()
-                    historyRecyclerView.isVisible = !adapter.isEmpty()
-                }
-                is PagingState.Loading -> {
-                    shimmerView.root.isVisible = adapter.isEmpty()
-                    refreshLayout.isVisible = true
-                    errorStateLayout.root.isVisible = false
-                    emptyStateLayout.root.isVisible = false
-                    historyRecyclerView.isVisible = !adapter.isEmpty()
-                }
-                is PagingState.Error -> {
-                    shimmerView.root.isVisible = false
-                    refreshLayout.isVisible = true
-                    errorStateLayout.root.isVisible = adapter.isEmpty()
-                    emptyStateLayout.root.isVisible = false
-                    historyRecyclerView.isVisible = !adapter.isEmpty()
-                }
-            }
-        }
-    }
-
-    override fun showHistory(history: List<HistoryItem>) {
-        with(binding.layoutHistoryList) {
-            adapter.setTransactions(history)
-            historyRecyclerView.invalidateItemDecorations()
-
-            val isHistoryEmpty = adapter.isEmpty()
-            emptyStateLayout.root.isVisible = isHistoryEmpty
-            historyRecyclerView.isVisible = !isHistoryEmpty
-        }
+        ) { _, _ -> binding.layoutHistoryList.loadHistory() }
     }
 
     override fun showBuyScreen(token: Token) {
@@ -149,15 +67,6 @@ class HistoryFragment :
                 Timber.e(IllegalArgumentException("Unsupported transaction type: $transaction"))
             }
         }
-    }
-
-    override fun showRefreshing(isRefreshing: Boolean) = with(binding.layoutHistoryList) {
-        refreshLayout.isRefreshing = isRefreshing
-        refreshLayoutProgressPlaceholder.isVisible = isRefreshing
-    }
-
-    override fun scrollToTop() {
-        binding.layoutHistoryList.historyRecyclerView.smoothScrollToPosition(0)
     }
 
     override fun openSellTransactionDetails(sellTransaction: SellTransactionViewDetails) {
