+++ app/src/main/java/org/p2p/wallet/history/ui/historylist/HistoryListViewContract.kt
@@ -5,12 +5,10 @@ import org.p2p.wallet.common.mvp.MvpPresenter
 import org.p2p.wallet.common.mvp.MvpView
 import org.p2p.wallet.common.ui.recycler.PagingState
 import org.p2p.wallet.history.model.HistoryItem
-import org.p2p.wallet.history.model.HistoryTransaction
 
 interface HistoryListViewContract {
     interface View : MvpView {
         fun showPagingState(state: PagingState)
-        fun openTransactionDetailsScreen(transaction: HistoryTransaction)
         fun showRefreshing(isRefreshing: Boolean)
         fun scrollToTop()
         fun showHistory(history: List<HistoryItem>)
@@ -20,6 +18,5 @@ interface HistoryListViewContract {
         fun loadHistory()
         fun refreshHistory()
         fun loadNextHistoryPage()
-        fun onItemClicked(historyItem: HistoryItem)
     }
 }
