+++ app/src/main/java/org/p2p/wallet/history/ui/token/TokenHistoryPresenter.kt
@@ -1,58 +1,28 @@
 package org.p2p.wallet.history.ui.token
 
-import androidx.lifecycle.LifecycleOwner
 import timber.log.Timber
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import org.p2p.core.token.Token
-import org.p2p.core.utils.merge
 import org.p2p.wallet.R
 import org.p2p.wallet.common.mvp.BasePresenter
-import org.p2p.wallet.common.ui.recycler.PagingState
 import org.p2p.wallet.common.ui.widget.actionbuttons.ActionButton
 import org.p2p.wallet.history.analytics.HistoryAnalytics
-import org.p2p.wallet.history.interactor.HistoryFetchListResult
-import org.p2p.wallet.history.interactor.HistoryInteractor
 import org.p2p.wallet.history.model.HistoryItem
 import org.p2p.wallet.history.model.HistoryTransaction
-import org.p2p.wallet.history.ui.history.HistorySellTransactionMapper
-import org.p2p.wallet.infrastructure.sell.HiddenSellTransactionsStorageContract
-import org.p2p.wallet.moonpay.model.SellTransaction
 import org.p2p.wallet.renbtc.interactor.RenBtcInteractor
 import org.p2p.wallet.rpc.interactor.TokenInteractor
-import org.p2p.wallet.sell.interactor.HistoryItemMapper
 import org.p2p.wallet.sell.ui.lock.SellTransactionViewDetails
 
 class TokenHistoryPresenter(
     private val token: Token.Active,
-    private val historyInteractor: HistoryInteractor,
-    private val hiddenSellTransactionsStorage: HiddenSellTransactionsStorageContract,
     private val historyAnalytics: HistoryAnalytics,
     private val renBtcInteractor: RenBtcInteractor,
     private val tokenInteractor: TokenInteractor,
-    private val sellTransactionsMapper: HistorySellTransactionMapper,
-    private val historyItemMapper: HistoryItemMapper,
 ) : BasePresenter<TokenHistoryContract.View>(), TokenHistoryContract.Presenter {
-
-    private object HistoryFetchFailure : Throwable(message = "Both transactions were not fetch due to errors")
-
-    private var blockChainTransactionsList = HistoryFetchListResult<HistoryTransaction>(mutableListOf())
-    private var sellTransactionsList = HistoryFetchListResult<SellTransaction>(mutableListOf())
-
-    private var pagingJob: Job? = null
-    private var loadingJob: Job? = null
-    private var refreshJob: Job? = null
-    private var paginationEnded: Boolean = false
 
     override fun attach(view: TokenHistoryContract.View) {
         super.attach(view)
         initialize()
-    }
-
-    override fun onResume(owner: LifecycleOwner) {
-        super.onResume(owner)
-        loadHistory()
     }
 
     private fun initialize() {
@@ -67,97 +37,6 @@ class TokenHistoryPresenter(
         }
 
         view?.showActionButtons(actionButtons)
-    }
-
-    override fun loadNextHistoryPage() {
-        if (paginationEnded) return
-
-        if (pagingJob?.isActive == true) {
-            return
-        }
-        pagingJob = launch {
-            view?.showPagingState(PagingState.Loading)
-            fetchHistory()
-        }
-    }
-
-    override fun retryLoad() {
-        paginationEnded = false
-        refreshJob?.cancel()
-
-        refreshJob = launch {
-            view?.showRefreshing(isRefreshing = true)
-            fetchHistory(isRefresh = true)
-            view?.showRefreshing(isRefreshing = false)
-            view?.scrollToTop()
-        }
-    }
-
-    override fun loadHistory() {
-        if (blockChainTransactionsList.hasFetchedItems() || sellTransactionsList.hasFetchedItems()) {
-            view?.showHistory(
-                historyItemMapper.fromDomainSell(sellTransactionsList.content).merge( // goes first
-                    historyItemMapper.fromDomainBlockchain(blockChainTransactionsList.content)
-                )
-            )
-            return
-        }
-        loadingJob = launch {
-            view?.showPagingState(PagingState.InitialLoading)
-            fetchHistory(isRefresh = true)
-        }
-    }
-
-    private suspend fun fetchHistory(isRefresh: Boolean = false) {
-        try {
-            if (isRefresh) {
-                blockChainTransactionsList.clearContent()
-                sellTransactionsList.clearContent()
-            }
-            if (token.isSOL) {
-                sellTransactionsList = fetchSellTransactions()
-            }
-            blockChainTransactionsList += fetchBlockChainTransactions(isRefresh)
-            if (blockChainTransactionsList.isFailed && sellTransactionsList.isFailed) {
-                view?.showPagingState(PagingState.Error(HistoryFetchFailure))
-                Timber.e(HistoryFetchFailure, "Error getting transaction history for token")
-            } else {
-                val sellTransactions = historyItemMapper.fromDomainSell(sellTransactionsList.content)
-                val blockchainTransactions = historyItemMapper.fromDomainBlockchain(blockChainTransactionsList.content)
-                view?.showHistory(
-                    sellTransactions.merge( // goes first
-                        blockchainTransactions
-                    )
-                )
-                view?.showPagingState(PagingState.Idle)
-            }
-        } catch (e: CancellationException) {
-            Timber.i(e, "Cancelled history next page load")
-        }
-    }
-
-    private suspend fun fetchBlockChainTransactions(isRefresh: Boolean) = try {
-        historyInteractor.loadTransactions(token.publicKey, isRefresh)
-            .toMutableList()
-            .let(::HistoryFetchListResult)
-    } catch (error: Throwable) {
-        Timber.e(error, "Error while loading blockchain transactions on history")
-        HistoryFetchListResult(isFailed = true)
-    }
-
-    private suspend fun fetchSellTransactions(): HistoryFetchListResult<SellTransaction> = try {
-        historyInteractor.getSellTransactions()?.let { sellTransactions ->
-            sellTransactionsMapper.map(sellTransactions)
-                .toMutableList()
-                .let(::HistoryFetchListResult)
-                .withContentFilter {
-                    !it.isCancelled() &&
-                        !hiddenSellTransactionsStorage.isTransactionHidden(it.transactionId)
-                }
-        } ?: HistoryFetchListResult(isFailed = true)
-    } catch (error: Throwable) {
-        Timber.e(error, "Error while loading Moonpay sell transactions on history")
-        HistoryFetchListResult(isFailed = true)
     }
 
     override fun onItemClicked(historyItem: HistoryItem) {
@@ -209,10 +88,5 @@ class TokenHistoryPresenter(
                 view?.showErrorMessage(e)
             }
         }
-    }
-
-    override fun onDestroy(owner: LifecycleOwner) {
-        super.onDestroy(owner)
-        loadingJob?.cancel()
     }
 }
